Initial implementation of deferred actions.

+++ include/eza/amd64/types.h
@@ -61,7 +61,7 @@ typedef int32_t pid_t;
 typedef uint16_t uid_t;
 typedef int32_t tid_t;
 typedef uint32_t usec_t;
-typedef int32_t status_t;
+typedef long status_t;
 typedef uint32_t shm_id_t;
 
 typedef uint64_t sigset_t;
